requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
ca-certificates=2019.5.15=1
certifi=2019.6.16=py27_1
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
icu=58.2=h211956c_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libuuid=1.0.3=h1bed415_2
libxml2=2.9.9=hea5a465_1
ncurses=6.1=he6710b0_1
openssl=1.1.1=h7b6447c_0
pip=19.2.2=py27_0
pyqt=4.11.4=py27_4
python=2.7.16=h8b3fad2_5
qt=4.8.7=2
readline=7.0=h7b6447c_5
setuptools=41.0.1=py27_0
sip=4.18=py27_0
sqlite=3.29.0=h7b6447c_0
tk=8.6.8=hbc83047_0
wheel=0.33.4=py27_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=h7b6447c_3
